AG-10217 - Add support for ordinalTimeScale time interval option
