+++ packages/ag-charts-community/src/sparklines-util.ts
@@ -13,7 +13,7 @@ export { bifurcate, extent, normalisedExtent, normalisedExtentWithMetadata } fro
 export { toFixed, isEqual as isNumberEqual } from './util/number';
 export { tickFormat } from './util/numberFormat';
 
-export { default as ticks, tickStep, range } from './util/ticks';
+export { default as ticks, tickStep, range, isDenseInterval } from './util/ticks';
 export { Color } from './util/color';
 export type { MeasuredLabel, PointLabelDatum } from './scene/util/labelPlacement';
 export * from './util/logger';
